add decorators selectors + make it works in bitbucket.org + improve display  (#20)
* Add default value for labels. To don't enable all labels by default.

fix #12

* feature: validation feature with custom decorator

Change function GetConventionalCommentPrefix to validate prefix with custom decorators. (by using regex).

Closes #13

* feature: fix validation search

Closes #13

* style: display tooltip in a better ways than title

* fix: Fix toolbar style present after click on submit

In first page of comment, when added a comment the toolbar, was steel in the interface. It was not removed. Next time to send a comment, a new toolbar will be created. But the old one, will not be removed. Fix this issues

* feat: Add a decorator selector

display a list a decorators checkbox that can be added after selecting labels.

fix #14

* feat: Add config to enable/disable decorators selector

add a config to enable/disabled configurators selectors + have a default label selector enabled.

fix #14

* feat: auto select current comment

If some comments with conventional comment were already added to the field. Already select it.

fix #14

* feat: add list of decorators in settings

Add a field to change lists of decorators (simple text-fields separated with ',').

fix #14

* :ambulance: fix: make it works to bitbucket.org

Change to make it works for both on premise bitbucket server and on Bitbucket.org cloud.

Closes #12

* :art: run prettier

* :bug: fix: fix bug when first click on label button + click on label buton

* :bulb: Remove comment (PR Review)

* :rewind: Remove unused files (PR Review)

---------

Co-authored-by: Emmanuel DE SAINT STEBAN <emmanuel.desaintsteban@atos.net>