Update CSS selectors to works with bitbucket.org
CSS selectors in selectors_cloud were updated to match new div structure in the editor toolbar. This change was made to improve compatibility with latest version of editor UI.

In app.js, a trailing semicolon was added for consistency and readability.

Function selectMatchingTextNode in utils.js was updated to better select item, and return boolean if selected Matching text or not.

fix #27