Fix bitbucket org  (#28)
* Update CSS selectors to works with bitbucket.org

CSS selectors in selectors_cloud were updated to match new div structure in the editor toolbar. This change was made to improve compatibility with latest version of editor UI.

In app.js, a trailing semicolon was added for consistency and readability.

Function selectMatchingTextNode in utils.js was updated to better select item, and return boolean if selected Matching text or not.

fix #27

* Update README and package.json details for extension update

The README has been updated to provide more comprehensive information about the enhanced features of the "Bitbucket-Conventional-Comments" Chrome extension. The update includes the addition of extension usage instructions and an enhanced description of the extension. Additional screenshots were added to illustrate the use of the extension on both Bitbucket.org and Bitbucket server.

The package.json file was updated to reflect the version bump to 2.1.0 and the addition of a new author to the project.

These changes help in providing clarity about the extension's usage, showcase its adaptability with both Bitbucket platforms, and acknowledge contributions to the project's development.

* :lipstick: Run prettier

* :rotating_light: fix issue with LIST_decorators for first users in app.js

Updated the listDecorator assignment in src/app.js to concatenate LIST_DECORATORS items using join, for correct error For a new user, there is an issue with decorators sometimes being persisted in state as an array instead of a string, causing .split to fail here..

* :bug: Fix issue when There's an issue when attempting to add a comment after something is already written in the box.

Refactor text cursor positioning and selection handling. this function now getting and selecting previousText, and past previous text with commitPrefix.

This commit addresses some inconsistencies in the cursor positioning and text selection logic in our contentEditable fields. In particular, the 'setCursorPosition' was renamed to 'selectPreviousText' to better depict its functionality. Instead, a new method 'selectPreviousText' was introduced, allowing for handling text preceding the cursor.

Additionally, a cleaner and more reliable method was introduced for selecting text nodes, avoiding possible mishaps with non-existing nodes. We changed 'decoratorsList' to 'decoratorList' for grammar and consistency throughout the code. The 'textNodeOffset' selector in 'selectors.js' was also updated to target 'p' elements instead of 'p strong' for a more general approach.

These changes provide a more robust handling of editable content, improving UX for text inputs and manipulation.

---------

Co-authored-by: Emmanuel DE SAINT STEBAN <emmanuel.desaintsteban@atos.net>